Move new classes to api & impl accordingly IllegalTreeStateException, DuplicateChildException, NodeCollection -> api IndexedNodeCollection -> impl
