+++ src/main/java/me/zeroeightsix/fiber/api/exception/DuplicateChildException.java
@@ -1,4 +1,4 @@
-package me.zeroeightsix.fiber.exception;
+package me.zeroeightsix.fiber.api.exception;
 
 public class DuplicateChildException extends IllegalTreeStateException {
     public DuplicateChildException(String message) {
