+++ src/main/java/me/zeroeightsix/fiber/api/exception/IllegalTreeStateException.java
@@ -1,7 +1,9 @@
-package me.zeroeightsix.fiber.exception;
+package me.zeroeightsix.fiber.api.exception;
+
+import me.zeroeightsix.fiber.api.tree.ConfigTree;
 
 /**
- * Signals that a {@link me.zeroeightsix.fiber.tree.ConfigTree} is not in an appropriate state for
+ * Signals that a {@link ConfigTree} is not in an appropriate state for
  * the requested operation.
  */
 public class IllegalTreeStateException extends IllegalStateException {
